+++ PatientScannerDemo/Services/CBOR.swift
@@ -61,47 +61,3 @@ struct CBOR {
     }
   }
 }
-
-extension SwiftCBOR.CBOR {
-  func toString() -> String {
-    switch self {
-    case let .byteString(val):
-      return String(describing: val)
-    case let .unsignedInt(val):
-      return "\(val)"
-    case let .negativeInt(val):
-      return "-\(val)"
-    case let .utf8String(val):
-      return "\"\(val)\""
-    case let .array(vals):
-      var s = ""
-      for val in vals {
-        s += (s.isEmpty ? "" : ", ") + val.toString()
-      }
-      return "[\(s)]"
-    case let .map(vals):
-      var s = ""
-      for pair in vals {
-        var key = pair.key.toString()
-        key = key.trimmingCharacters(in: ["\""])
-        key = "\"\(key)\""
-        s += (s.isEmpty ? "" : ", ") + "\(key): \(pair.value.toString())"
-      }
-      return "{\(s)}"
-    case let .boolean(val):
-      return String(describing: val)
-    case .null:
-      return "null"
-    case .undefined:
-      return "null"
-    case let .float(val):
-      return "\(val)"
-    case let .double(val):
-      return "\(val)"
-    case let .date(val):
-      return "\"\(val.isoString)\""
-    default:
-      return "\"unsupported data\""
-    }
-  }
-}
